Export Resources tests
